+++ Tests/MultipasteCoreTests/main.swift
@@ -4,5 +4,7 @@ ClipboardItemTests.registerAll()
 HistoryStoreTests.registerAll()
 PreferencesTests.registerAll()
 SnippetMatcherTests.registerAll()
+SemanticVersionTests.registerAll()
+UpdateCheckerTests.registerAll()
 
 exit(Int32(runAllTests()))